Events are now adding to the db, but they're not refreshing. I also changed the login page to look a little nicer.

+++ app/src/main/java/edu/byu/stringcheese/presenttime/database/Event.java
@@ -35,7 +35,7 @@ public class Event {
         this.items = new HashMap<>();
     }
 
-    public String addItem(String name, int cost, String store, int imageID) {
+    public String addItem(String name, double cost, String store, int imageID) {
         //add item
         Firebase items = FirebaseDatabase.ref.child("items");
         Firebase newItem = items.push();
